network_monitor/enterprise_monitor.py
from datetime import datetime
import logging
from collections import defaultdict
import requests
import matplotlib.pyplot as plt
import numpy as np
import yara
import dns.resolver
import ssl
import subprocess
from scapy.all import *
from reportlab.pdfgen import canvas
from elasticsearch import Elasticsearch

# Define analyzer classes first
class DNSAnalyzer:
    def __init__(self):
        self.suspicious_domains = set()
        self.dns_cache = {}
        
    def analyze_dns_traffic(self, domain):
        # Check domain reputation
        reputation = self.check_domain_reputation(domain)
        
        # Detect DNS tunneling
        if self.detect_dns_tunneling(domain):
            self.alert_dns_tunnel(domain)
            
        # Check for DGA patterns
        if self.is_dga_domain(domain):
            self.alert_dga_detected(domain)
            
        return reputation
        
    def detect_dns_tunneling(self, domain):
        # Implement DNS tunneling detection logic
        pass

class SSLAnalyzer:
    def __init__(self):
        self.ssl_fingerprints = set()
        
    def analyze_ssl_connection(self, connection):
        # Get JA3 fingerprint
        ja3_hash = self.get_ja3_fingerprint(connection)
        
        # Validate certificate
        cert_status = self.validate_certificate(connection)
        
        # Check TLS version
        tls_version = self.check_tls_version(connection)
        
        # Analyze cipher suites
        cipher_analysis = self.analyze_cipher_suites(connection)
        
        return {
            'ja3_hash': ja3_hash,
            'cert_status': cert_status,
            'tls_version': tls_version,
            'cipher_analysis': cipher_analysis
        }

class ProtocolAnalyzer:
    def __init__(self):
        self.protocol_stats = defaultdict(int)
        
    def analyze_protocol(self, packet):
        if TCP in packet:
            port = packet[TCP].dport
            if port == 80:
                return self.analyze_http(packet)
            elif port == 443:
                return self.analyze_https(packet)
            elif port == 21:
                return self.analyze_ftp(packet)
            elif port == 25:
                return self.analyze_smtp(packet)
        
        return self.analyze_custom_protocol(packet)

class ThreatHunter:
    def __init__(self):
        self.yara_rules = {}
        self.ioc_database = set()
        self.mitre_mappings = {}
        
    def hunt_threats(self, data):
        # Apply YARA rules
        yara_matches = self.scan_with_yara(data)
        
        # Check IOCs
        ioc_matches = self.check_iocs(data)
        
        # Map to MITRE ATT&CK
        mitre_techniques = self.map_to_mitre(yara_matches + ioc_matches)
        
        return {
            'yara_matches': yara_matches,
            'ioc_matches': ioc_matches,
            'mitre_techniques': mitre_techniques
        }

class ResourceMonitor:
    def __init__(self):
        self.resource_history = defaultdict(list)
        
    def monitor_resources(self):
        cpu_usage = psutil.cpu_percent(interval=1, percpu=True)
        memory_usage = psutil.virtual_memory()
        disk_io = psutil.disk_io_counters()
        network_buffers = self.get_network_buffers()
        
        self.analyze_resource_patterns(cpu_usage, memory_usage, disk_io)
        return self.generate_resource_report()

# Main monitor class
class EnterpriseNetworkMonitor:
    def __init__(self):
        self.setup_components()
        self.setup_elasticsearch()
        self.load_yara_rules()
        self.initialize_dashboard()
        
    def setup_components(self):
        self.dns_analyzer = DNSAnalyzer()
        self.ssl_analyzer = SSLAnalyzer()
        self.protocol_analyzer = ProtocolAnalyzer()
        self.threat_hunter = ThreatHunter()
        self.resource_monitor = ResourceMonitor()
        self.rule_engine = RuleEngine()
        self.report_generator = ReportGenerator()

    # ... rest of EnterpriseNetworkMonitor implementation ...

if __name__ == "__main__":
    monitor = EnterpriseNetworkMonitor()
    monitor.monitor_network()