network_monitor/requirements.txt
psutil==5.9.5
matplotlib==3.7.1
requests==2.31.0
scapy==2.5.0
yara-python==4.3.1
folium==0.14.0
numpy==1.24.3
